Refactor graphs as much as possible.
